Move issueReportSchema
