+++ src/types/issueReportTypes.ts
@@ -1,5 +1,3 @@
-import { z } from 'zod'
-
 export type DefaultField = 'Workflow' | 'Logs' | 'SystemStats' | 'Settings'
 
 export interface ReportField {
@@ -51,15 +49,3 @@ export interface IssueReportPanelProps {
    */
   title?: string
 }
-
-const checkboxField = z.boolean().optional()
-export const issueReportSchema = z
-  .object({
-    contactInfo: z.string().email().max(320).optional().or(z.literal('')),
-    details: z.string().max(5_000).optional()
-  })
-  .catchall(checkboxField)
-  .refine((data) => Object.values(data).some((value) => value), {
-    path: ['details']
-  })
-export type IssueReportFormData = z.infer<typeof issueReportSchema>
